requirements.txt
requests
flask
flask_cors
flask-limiter
python-dotenv
gunicorn
pymemcache
